refactor(miniapp): code refactor of native/plugin components
