+++ packages/rax-miniapp-babel-plugins/src/plugins/babel-plugin-handle-plugin-component.js
@@ -1,0 +1,34 @@
+const { collectComponentAttr, collectUsings, scanSlot } = require('../utils/handleComponentAST');
+const MINIAPP_PLUGIN_COMPONENTS_REG = /^plugin\:\/\//;
+
+module.exports = function visitor(
+  { types: t },
+  { usingPlugins }
+) {
+  // Collect imported dependencies
+  const pluginComponents = {};
+  const pluginComponentsNameMap = new Map();
+  const scanedFileMap = {};
+
+  return {
+    visitor: {
+      ImportDeclaration: {
+        enter(path, { filename }) {
+          const { source } = path.node;
+          if (t.isStringLiteral(source) && MINIAPP_PLUGIN_COMPONENTS_REG.test(source.value)) {
+            if (!scanedFileMap[filename]) {
+              scanedFileMap[filename] = true;
+              path.parentPath.traverse({
+                JSXOpeningElement: collectComponentAttr(pluginComponents)
+              });
+            }
+            collectUsings(path, pluginComponents, pluginComponentsNameMap, usingPlugins, source.value);
+          }
+        },
+      },
+      JSXOpeningElement: {
+        exit: scanSlot(pluginComponentsNameMap, usingPlugins)
+      },
+    },
+  };
+};
